+++ app/ui/SingelProduct.tsx
@@ -1,9 +1,5 @@
-"use client";
-
-import { Box, Grid, Snackbar, Typography } from "@mui/material";
+import { Box, Grid, Typography } from "@mui/material";
 import { Product } from "@prisma/client";
-import { useState } from "react";
-import { useCart } from "../context/CartContext";
 import AddToCartButton from "./AddToCartButton";
 
 type Props = {
@@ -11,21 +7,6 @@ type Props = {
 };
 
 export default function SingleProduct({ product }: Props) {
-  const [openSnackbar, setOpenSnackbar] = useState(false); // Tillstånd för att visa snackbar
-  const [snackbarMessage, setSnackbarMessage] = useState(""); // Tillstånd för meddelandet i snackbar
-
-  const { addToCart } = useCart();
-
-  const handleAddToCart = () => {
-    addToCart(product);
-    setSnackbarMessage(`${product.title} har lagts till i kundvagnen`); // Ange meddelandet för snackbar
-    setOpenSnackbar(true); // Visa snackbar
-  };
-
-  const handleCloseSnackbar = () => {
-    setOpenSnackbar(false); // Stäng snackbar
-  };
-
   return (
     <main>
       <Grid container spacing={2}>
@@ -58,14 +39,6 @@ export default function SingleProduct({ product }: Props) {
           </Box>
         </Grid>
       </Grid>
-      {/* Snackbar för att visa meddelande när en produkt läggs till i kundvagnen */}
-      <Snackbar
-        anchorOrigin={{ vertical: "top", horizontal: "right" }}
-        open={openSnackbar}
-        autoHideDuration={1000}
-        onClose={handleCloseSnackbar}
-        message={snackbarMessage}
-      />
     </main>
   );
 }
